Make the modulemap for all platforms generated by gn
By making it generated, we can allow users to override the sysroot
directory. This is relevant for ChromeOS, which uses cros-sdk to obtain
a toolchain.

It is also relevant for windows, as each new version of the windows
sysroot has its hash in the path, and this will allow the modulemap to
work between updates.

Bug: 442479236
Change-Id: I6a6a6964e8f34b12fae20e095580a59bbca8b863
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6904866
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Matt Stark <msta@google.com>
Cr-Commit-Position: refs/heads/main@{#1510094}
NOKEYCHECK=True
GitOrigin-RevId: 6ff7bdb3d5e40631a47cc100311675d3204487da